requirements.txt
# ChAnki v2 - Production Dependencies
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# AI & NLP
google-generativeai==0.3.2
jieba==0.42.1
pypinyin==0.51.0

# HTTP & API
requests==2.31.0
httpx==0.26.0

# Audio Generation
edge-tts==6.1.9

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
